unify SPI tests under one test suite
