--- hil-test/src/bin/spi_half_duplex_write_psram.rs
@@ -1,186 +1,0 @@
-//! SPI Half Duplex Write Test
-
-// FIXME: Only validated on the ESP32-S3 HIL bench, other PSRAM chips pick up spurious PCNT pulses
-// there.
-//% CHIP_FILTER: esp32s3
-//% FEATURES: unstable esp-alloc
-
-#![no_std]
-#![no_main]
-use defmt::error;
-use esp_alloc as _;
-use esp_hal::{
-    Blocking,
-    dma::ExternalBurstConfig,
-    dma_rx_buffer,
-    gpio::{Flex, interconnect::InputSignal},
-    pcnt::{Pcnt, channel::EdgeMode, unit::Unit},
-    spi::{
-        Mode,
-        master::{Address, Command, Config, DataMode, Spi, SpiDma},
-    },
-    time::Rate,
-};
-use hil_test as _;
-extern crate alloc;
-
-macro_rules! dma_alloc_tx_buffer {
-    ($size:expr, $align:expr) => {{
-        let layout = core::alloc::Layout::from_size_align($size, $align as usize).unwrap();
-        let buffer = unsafe {
-            let ptr = alloc::alloc::alloc(layout);
-            if ptr.is_null() {
-                error!("dma_alloc_buffer: alloc failed");
-                alloc::alloc::handle_alloc_error(layout);
-            }
-            core::slice::from_raw_parts_mut(ptr, $size)
-        };
-
-        const DMA_CHUNK_SIZE: usize = 4096 - $align as usize;
-        let descriptors = esp_hal::dma_descriptors_impl!($size, DMA_CHUNK_SIZE);
-        esp_hal::dma::DmaTxBuf::new_with_config(
-            descriptors,
-            unsafe { esp_hal::dma::aligned::DmaAlignedMut::new_unchecked(buffer) },
-            $align,
-        )
-    }};
-}
-
-struct Context {
-    spi: SpiDma<'static, Blocking>,
-    pcnt_unit: Unit<'static, 0>,
-    pcnt_source: InputSignal<'static>,
-}
-
-#[embedded_test::tests(default_timeout = 3)]
-mod tests {
-
-    use super::*;
-
-    #[init]
-    fn init() -> Context {
-        let peripherals = esp_hal::init(
-            esp_hal::Config::default().with_cpu_clock(esp_hal::clock::CpuClock::max()),
-        );
-        esp_alloc::psram_allocator!(peripherals.PSRAM, esp_hal::psram);
-
-        let sclk = peripherals.GPIO0;
-        let (mosi, _) = hil_test::common_test_pins!(peripherals);
-
-        let pcnt = Pcnt::new(peripherals.PCNT);
-
-        let dma_channel = cfg_select! {
-            spi_master_dma_engine = "SPI_DMA" => peripherals.DMA_SPI2,
-            spi_master_dma_engine = "AHB_GDMA" => peripherals.DMA_CH0,
-            spi_master_dma_engine = "AXI_GDMA" => peripherals.DMA_AXI_CH0,
-        };
-
-        let mut mosi = Flex::new(mosi);
-        mosi.set_input_enable(true);
-        mosi.set_output_enable(true);
-        let mosi_loopback = mosi.peripheral_input();
-
-        let spi = Spi::new(
-            peripherals.SPI2,
-            Config::default()
-                .with_frequency(Rate::from_khz(100))
-                .with_mode(Mode::_0),
-        )
-        .unwrap()
-        .with_sck(sclk)
-        .with_sio0(mosi)
-        .with_dma(dma_channel);
-
-        Context {
-            spi,
-            pcnt_unit: pcnt.unit0,
-            pcnt_source: mosi_loopback,
-        }
-    }
-
-    #[test]
-    fn test_spi_writes_are_correctly_by_pcnt(ctx: Context) {
-        const DMA_BUFFER_SIZE: usize = 4;
-        const DMA_ALIGNMENT: ExternalBurstConfig = ExternalBurstConfig::Size32;
-        let mut dma_tx_buf = dma_alloc_tx_buffer!(DMA_BUFFER_SIZE, DMA_ALIGNMENT).unwrap();
-
-        let unit = ctx.pcnt_unit;
-        let mut spi = ctx.spi;
-
-        unit.channel0.set_edge_signal(ctx.pcnt_source);
-        unit.channel0
-            .set_input_mode(EdgeMode::Hold, EdgeMode::Increment);
-
-        // Fill the buffer where each byte has 3 pos edges.
-        dma_tx_buf.fill(&[0b0110_1010; DMA_BUFFER_SIZE]);
-        let transfer = spi
-            .half_duplex_write_buffer(
-                DataMode::SingleTwoDataLines,
-                Command::None,
-                Address::None,
-                0,
-                dma_tx_buf.len(),
-                dma_tx_buf,
-            )
-            .map_err(|e| e.0)
-            .unwrap();
-        (spi, dma_tx_buf) = transfer.wait();
-
-        assert_eq!(unit.value(), (3 * DMA_BUFFER_SIZE) as _);
-
-        let transfer = spi
-            .half_duplex_write_buffer(
-                DataMode::SingleTwoDataLines,
-                Command::None,
-                Address::None,
-                0,
-                dma_tx_buf.len(),
-                dma_tx_buf,
-            )
-            .map_err(|e| e.0)
-            .unwrap();
-        transfer.wait();
-
-        assert_eq!(unit.value(), (6 * DMA_BUFFER_SIZE) as _);
-    }
-
-    #[test]
-    fn test_spidmabus_writes_are_correctly_by_pcnt(ctx: Context) {
-        const DMA_BUFFER_SIZE: usize = 4;
-        const DMA_ALIGNMENT: ExternalBurstConfig = ExternalBurstConfig::Size32; // matches dcache line size
-
-        let dma_tx_buf = dma_alloc_tx_buffer!(DMA_BUFFER_SIZE, DMA_ALIGNMENT).unwrap();
-        let dma_rx_buf = dma_rx_buffer!(1).unwrap();
-
-        let unit = ctx.pcnt_unit;
-        let mut spi = ctx.spi.with_buffers(dma_rx_buf, dma_tx_buf);
-
-        unit.channel0.set_edge_signal(ctx.pcnt_source);
-        unit.channel0
-            .set_input_mode(EdgeMode::Hold, EdgeMode::Increment);
-
-        let buffer = [0b0110_1010; DMA_BUFFER_SIZE];
-        // Write the buffer where each byte has 3 pos edges.
-        spi.half_duplex_write(
-            DataMode::SingleTwoDataLines,
-            Command::None,
-            Address::None,
-            0,
-            &buffer,
-        )
-        .unwrap();
-
-        assert_eq!(unit.value(), (3 * DMA_BUFFER_SIZE) as _);
-
-        spi.half_duplex_write(
-            DataMode::SingleTwoDataLines,
-            Command::None,
-            Address::None,
-            0,
-            &buffer,
-        )
-        .unwrap();
-
-        assert_eq!(unit.value(), (6 * DMA_BUFFER_SIZE) as _);
-    }
-}
